Added Sensor support to Sphere constraints.  Spring sensors are not needed, due to discovery that you'd need to know the spring number, so you may as well know the node numbers and do your own sensor related calculations.
git-svn-id: svn://10.0.50.55/puffbomb/trunk@1351 d14347b8-48fd-0310-9e84-97907d434d85

+++ Code/Engine2D/Physics/Linkage/NodeAnchor.h
@@ -28,7 +28,7 @@ public:
 	
 
 	// Object number and Index numbers used //
-	unsigned int Object, Index;
+	size_t Object, Index;
 		
 public:
 	inline void Step( std::vector< cDynamicObject >& Component ) {
